@@ -1,24 +1,24 @@
-import { injectable, inject } from 'inversify';
-import { Context } from 'telegraf';
-import { TYPES } from '../types';
-import { LoggerService } from './logger.service';
-import { IErrorService } from '../interfaces';
+import { injectable, inject } from "inversify";
+import { Context } from "telegraf";
+import { TYPES } from "../types";
+import { LoggerService } from "./logger.service";
+import { IErrorService } from "../interfaces";
 
 export enum ErrorType {
-  GENERAL = 'general',
-  NETWORK = 'network',
-  VALIDATION = 'validation',
-  SUBSCRIPTION = 'subscription',
-  FILE_PROCESSING = 'file_processing',
-  AUTHENTICATION = 'authentication',
-  RATE_LIMIT = 'rate_limit',
-  EXTERNAL_API = 'external_api'
+  GENERAL = "general",
+  NETWORK = "network",
+  VALIDATION = "validation",
+  SUBSCRIPTION = "subscription",
+  FILE_PROCESSING = "file_processing",
+  AUTHENTICATION = "authentication",
+  RATE_LIMIT = "rate_limit",
+  EXTERNAL_API = "external_api",
 }
 
 export interface ErrorContext {
   userId?: string;
   action?: string;
-  details?: any;
+  details?: unknown;
   showDeveloperContact?: boolean;
   rollback?: () => Promise<void>;
 }
@@ -27,9 +27,7 @@ export interface ErrorContext {
 export class ErrorService implements IErrorService {
   private loggerService: LoggerService;
 
-  constructor(
-    @inject(TYPES.LoggerService) loggerService: LoggerService
-  ) {
+  constructor(@inject(TYPES.LoggerService) loggerService: LoggerService) {
     this.loggerService = loggerService;
   }
 
@@ -41,35 +39,37 @@ export class ErrorService implements IErrorService {
     ctx: Context,
     error: Error | unknown,
     errorType: ErrorType = ErrorType.GENERAL,
-    errorContext?: ErrorContext
+    errorContext?: ErrorContext,
   ): Promise<void> {
     const errorMessage = error instanceof Error ? error.message : String(error);
     const stack = error instanceof Error ? error.stack : undefined;
-    
+
     // Execute rollback if provided
     if (errorContext?.rollback) {
       try {
-        console.log(`🔄 [ERROR ROLLBACK] Executing rollback for user ${errorContext.userId}`);
+        console.log(
+          `🔄 [ERROR ROLLBACK] Executing rollback for user ${errorContext.userId}`,
+        );
         await errorContext.rollback();
       } catch (rollbackError) {
-        this.loggerService.error('Failed to execute rollback', {
+        this.loggerService.error("Failed to execute rollback", {
           originalError: errorMessage,
-          rollbackError: rollbackError instanceof Error ? rollbackError.message : String(rollbackError),
-          userId: errorContext?.userId
+          rollbackError:
+            rollbackError instanceof Error
+              ? rollbackError.message
+              : String(rollbackError),
+          userId: errorContext?.userId,
         });
       }
     }
-    
+
     // Log the error with context
-    this.loggerService.error(
-      `[${errorType.toUpperCase()}] ${errorMessage}`,
-      {
-        userId: errorContext?.userId || ctx.from?.id,
-        action: errorContext?.action,
-        details: errorContext?.details,
-        stack
-      }
-    );
+    this.loggerService.error(`[${errorType.toUpperCase()}] ${errorMessage}`, {
+      userId: errorContext?.userId || ctx.from?.id,
+      action: errorContext?.action,
+      details: errorContext?.details,
+      stack,
+    });
 
     // Send user-friendly error message
     await this.sendErrorMessage(ctx, errorType, errorContext);
@@ -81,25 +81,27 @@ export class ErrorService implements IErrorService {
   async sendErrorMessage(
     ctx: Context,
     errorType: ErrorType = ErrorType.GENERAL,
-    errorContext?: ErrorContext
+    errorContext?: ErrorContext,
   ): Promise<void> {
     try {
       const baseMessage = this.getErrorMessage(errorType);
-      const developerContact = errorContext?.showDeveloperContact !== false 
-        ? 'If this problem persists, please contact support.'
-        : '';
-      
-      const fullMessage = developerContact 
+      const developerContact =
+        errorContext?.showDeveloperContact !== false
+          ? "If this problem persists, please contact support."
+          : "";
+
+      const fullMessage = developerContact
         ? `${baseMessage}\n\n${developerContact}`
         : baseMessage;
 
-      await ctx.reply(fullMessage, { parse_mode: 'Markdown' });
+      await ctx.reply(fullMessage, { parse_mode: "Markdown" });
     } catch (replyError) {
       // Fallback if reply fails
-      this.loggerService.error('Failed to send error message to user', {
+      this.loggerService.error("Failed to send error message to user", {
         originalError: errorType,
-        replyError: replyError instanceof Error ? replyError.message : String(replyError),
-        userId: ctx.from?.id
+        replyError:
+          replyError instanceof Error ? replyError.message : String(replyError),
+        userId: ctx.from?.id,
       });
     }
   }
@@ -109,16 +111,21 @@ export class ErrorService implements IErrorService {
    */
   private getErrorMessage(errorType: ErrorType): string {
     const errorMessages = {
-      [ErrorType.GENERAL]: 'Something went wrong. Please try again.',
-      [ErrorType.NETWORK]: 'Network error. Please check your connection and try again.',
-      [ErrorType.VALIDATION]: 'Invalid input. Please check your request and try again.',
-      [ErrorType.SUBSCRIPTION]: 'Subscription error. Please check your subscription status.',
-      [ErrorType.FILE_PROCESSING]: 'File processing error. Please try again.',
-      [ErrorType.AUTHENTICATION]: 'Authentication error. Please try again.',
-      [ErrorType.RATE_LIMIT]: 'Rate limit exceeded. Please wait a moment and try again.',
-      [ErrorType.EXTERNAL_API]: 'External service error. Please try again later.'
+      [ErrorType.GENERAL]: "Something went wrong. Please try again.",
+      [ErrorType.NETWORK]:
+        "Network error. Please check your connection and try again.",
+      [ErrorType.VALIDATION]:
+        "Invalid input. Please check your request and try again.",
+      [ErrorType.SUBSCRIPTION]:
+        "Subscription error. Please check your subscription status.",
+      [ErrorType.FILE_PROCESSING]: "File processing error. Please try again.",
+      [ErrorType.AUTHENTICATION]: "Authentication error. Please try again.",
+      [ErrorType.RATE_LIMIT]:
+        "Rate limit exceeded. Please wait a moment and try again.",
+      [ErrorType.EXTERNAL_API]:
+        "External service error. Please try again later.",
     };
-    
+
     return errorMessages[errorType] || errorMessages[ErrorType.GENERAL];
   }
 
@@ -130,25 +137,30 @@ export class ErrorService implements IErrorService {
     ctx: Context,
     error: Error | unknown,
     errorType: ErrorType = ErrorType.GENERAL,
-    errorContext?: ErrorContext
+    errorContext?: ErrorContext,
   ): Promise<void> {
     const errorMessage = error instanceof Error ? error.message : String(error);
     const stack = error instanceof Error ? error.stack : undefined;
-    
+
     // Execute rollback if provided
     if (errorContext?.rollback) {
       try {
-        console.log(`🔄 [CALLBACK ERROR ROLLBACK] Executing rollback for user ${errorContext.userId}`);
+        console.log(
+          `🔄 [CALLBACK ERROR ROLLBACK] Executing rollback for user ${errorContext.userId}`,
+        );
         await errorContext.rollback();
       } catch (rollbackError) {
-        this.loggerService.error('Failed to execute callback rollback', {
+        this.loggerService.error("Failed to execute callback rollback", {
           originalError: errorMessage,
-          rollbackError: rollbackError instanceof Error ? rollbackError.message : String(rollbackError),
-          userId: errorContext?.userId
+          rollbackError:
+            rollbackError instanceof Error
+              ? rollbackError.message
+              : String(rollbackError),
+          userId: errorContext?.userId,
         });
       }
     }
-    
+
     // Log the error
     this.loggerService.error(
       `[CALLBACK_${errorType.toUpperCase()}] ${errorMessage}`,
@@ -156,23 +168,24 @@ export class ErrorService implements IErrorService {
         userId: errorContext?.userId || ctx.from?.id,
         action: errorContext?.action,
         details: errorContext?.details,
-        stack
-      }
+        stack,
+      },
     );
 
     // Answer callback query with error and send message
     try {
-      if ('answerCbQuery' in ctx) {
-        const shortError = 'Error occurred';
+      if ("answerCbQuery" in ctx) {
+        const shortError = "Error occurred";
         await ctx.answerCbQuery(shortError);
       }
-      
+
       await this.sendErrorMessage(ctx, errorType, errorContext);
     } catch (replyError) {
-      this.loggerService.error('Failed to handle callback error', {
+      this.loggerService.error("Failed to handle callback error", {
         originalError: errorType,
-        replyError: replyError instanceof Error ? replyError.message : String(replyError),
-        userId: ctx.from?.id
+        replyError:
+          replyError instanceof Error ? replyError.message : String(replyError),
+        userId: ctx.from?.id,
       });
     }
   }
@@ -185,11 +198,11 @@ export class ErrorService implements IErrorService {
     messageId: number,
     error: Error | unknown,
     errorType: ErrorType = ErrorType.GENERAL,
-    errorContext?: ErrorContext
+    errorContext?: ErrorContext,
   ): Promise<void> {
     const errorMessage = error instanceof Error ? error.message : String(error);
     const stack = error instanceof Error ? error.stack : undefined;
-    
+
     // Log the error
     this.loggerService.error(
       `[MESSAGE_EDIT_${errorType.toUpperCase()}] ${errorMessage}`,
@@ -198,18 +211,19 @@ export class ErrorService implements IErrorService {
         action: errorContext?.action,
         messageId,
         details: errorContext?.details,
-        stack
-      }
+        stack,
+      },
     );
 
     // Try to edit the message with error, fallback to new message
     try {
       const baseMessage = this.getErrorMessage(errorType);
-      const developerContact = errorContext?.showDeveloperContact !== false 
-        ? 'If this problem persists, please contact support.'
-        : '';
-      
-      const fullMessage = developerContact 
+      const developerContact =
+        errorContext?.showDeveloperContact !== false
+          ? "If this problem persists, please contact support."
+          : "";
+
+      const fullMessage = developerContact
         ? `${baseMessage}\n\n${developerContact}`
         : baseMessage;
 
@@ -220,19 +234,20 @@ export class ErrorService implements IErrorService {
             messageId,
             undefined,
             fullMessage,
-            { parse_mode: 'Markdown' }
+            { parse_mode: "Markdown" },
           );
-        } catch (editError) {
+        } catch {
           // If editing fails, send a new message
-          await ctx.reply(fullMessage, { parse_mode: 'Markdown' });
+          await ctx.reply(fullMessage, { parse_mode: "Markdown" });
         }
       }
     } catch (replyError) {
-      this.loggerService.error('Failed to handle message edit error', {
+      this.loggerService.error("Failed to handle message edit error", {
         originalError: errorType,
-        replyError: replyError instanceof Error ? replyError.message : String(replyError),
+        replyError:
+          replyError instanceof Error ? replyError.message : String(replyError),
         userId: ctx.from?.id,
-        messageId
+        messageId,
       });
     }
   }
@@ -243,16 +258,16 @@ export class ErrorService implements IErrorService {
   createErrorContext(
     userId?: string,
     action?: string,
-    details?: any,
+    details?: unknown,
     showDeveloperContact: boolean = true,
-    rollback?: () => Promise<void>
+    rollback?: () => Promise<void>,
   ): ErrorContext {
     return {
       userId,
       action,
       details,
       showDeveloperContact,
-      rollback
+      rollback,
     };
   }
 
@@ -262,19 +277,16 @@ export class ErrorService implements IErrorService {
   logError(
     error: Error | unknown,
     errorType: ErrorType = ErrorType.GENERAL,
-    errorContext?: ErrorContext
+    errorContext?: ErrorContext,
   ): void {
     const errorMessage = error instanceof Error ? error.message : String(error);
     const stack = error instanceof Error ? error.stack : undefined;
-    
-    this.loggerService.error(
-      `[${errorType.toUpperCase()}] ${errorMessage}`,
-      {
-        userId: errorContext?.userId,
-        action: errorContext?.action,
-        details: errorContext?.details,
-        stack
-      }
-    );
+
+    this.loggerService.error(`[${errorType.toUpperCase()}] ${errorMessage}`, {
+      userId: errorContext?.userId,
+      action: errorContext?.action,
+      details: errorContext?.details,
+      stack,
+    });
   }
-} 
+}
