@@ -1,20 +1,27 @@
-import { injectable, inject } from 'inversify';
-import { Connection, Keypair, PublicKey, Transaction, SystemProgram, LAMPORTS_PER_SOL } from '@solana/web3.js';
-import { TYPES } from '../types';
-import { SupabaseService } from './supabase.service';
-import { solanaRpcUrl } from '../config';
-import { VolumeOrder, VolumeTask, TaskStatus, IVolumeEngineService, OrderStatus } from '../interfaces';
-import { v4 as uuidv4 } from 'uuid';
+import { injectable, inject } from "inversify";
+import { Connection, Keypair } from "@solana/web3.js";
+import { TYPES } from "../types";
+import { SupabaseService } from "./supabase.service";
+import { solanaRpcUrl } from "../config";
+import {
+  VolumeOrder,
+  VolumeTask,
+  TaskStatus,
+  IVolumeEngineService,
+  OrderStatus,
+} from "../interfaces";
+// import { v4 as uuidv4 } from 'uuid'; // Unused
 
 @injectable()
 export class VolumeEngineService implements IVolumeEngineService {
   private connection: Connection;
   private supabaseService: SupabaseService;
 
-  constructor(
-    @inject(TYPES.SupabaseService) supabaseService: SupabaseService
-  ) {
-    this.connection = new Connection(solanaRpcUrl, 'confirmed');
+  constructor(@inject(TYPES.SupabaseService) supabaseService: SupabaseService) {
+    this.connection = new Connection(
+      solanaRpcUrl || "https://api.mainnet-beta.solana.com",
+      "confirmed",
+    );
     this.supabaseService = supabaseService;
   }
 
@@ -22,33 +29,40 @@ export class VolumeEngineService implements IVolumeEngineService {
     try {
       const order = await this.supabaseService.getVolumeOrder(orderId);
       if (!order) {
-        throw new Error('Order not found');
+        throw new Error("Order not found");
       }
 
       if (order.status !== OrderStatus.PAYMENT_CONFIRMED) {
-        throw new Error('Order payment not confirmed');
+        throw new Error("Order payment not confirmed");
       }
 
       // Update order status to running
       await this.supabaseService.updateVolumeOrder(orderId, {
         status: OrderStatus.RUNNING,
-        started_at: new Date().toISOString()
+        started_at: new Date().toISOString(),
       });
 
       // Create volume tasks
       const tasks = await this.createVolumeTasks(order);
-      
-      console.log(`✅ Started volume generation for order ${orderId} with ${tasks.length} tasks`);
-      console.log(`📊 Tasks will be processed by the volume-processor cron job every 5 minutes`);
+
+      console.log(
+        `✅ Started volume generation for order ${orderId} with ${tasks.length} tasks`,
+      );
+      console.log(
+        `📊 Tasks will be processed by the volume-processor cron job every 5 minutes`,
+      );
     } catch (error) {
-      console.error(`❌ Error starting volume generation for order ${orderId}:`, error);
-      
+      console.error(
+        `❌ Error starting volume generation for order ${orderId}:`,
+        error,
+      );
+
       // Update order status to failed
       await this.supabaseService.updateVolumeOrder(orderId, {
         status: OrderStatus.FAILED,
-        completed_at: new Date().toISOString()
+        completed_at: new Date().toISOString(),
       });
-      
+
       throw error;
     }
   }
@@ -58,7 +72,7 @@ export class VolumeEngineService implements IVolumeEngineService {
       // Update order status
       await this.supabaseService.updateVolumeOrder(orderId, {
         status: OrderStatus.COMPLETED,
-        completed_at: new Date().toISOString()
+        completed_at: new Date().toISOString(),
       });
 
       // Update all running tasks to completed
@@ -66,14 +80,17 @@ export class VolumeEngineService implements IVolumeEngineService {
       for (const task of tasks) {
         if (task.status === TaskStatus.RUNNING) {
           await this.supabaseService.updateVolumeTask(task.id, {
-            status: TaskStatus.COMPLETED
+            status: TaskStatus.COMPLETED,
           });
         }
       }
 
       console.log(`✅ Stopped volume generation for order ${orderId}`);
     } catch (error) {
-      console.error(`❌ Error stopping volume generation for order ${orderId}:`, error);
+      console.error(
+        `❌ Error stopping volume generation for order ${orderId}:`,
+        error,
+      );
       throw error;
     }
   }
@@ -87,20 +104,22 @@ export class VolumeEngineService implements IVolumeEngineService {
   }
 
   private async createVolumeTasks(order: VolumeOrder): Promise<VolumeTask[]> {
-    const tasks: Omit<VolumeTask, 'id' | 'created_at' | 'updated_at'>[] = [];
-    
+    const tasks: Omit<VolumeTask, "id" | "created_at" | "updated_at">[] = [];
+
     // Calculate volume per task
     const volumePerTask = order.volume_target / order.tasks_count;
-    
+
     // Calculate interval between cycles (not tasks)
     // Each task should complete its 5 cycles over the duration
     const totalCycles = order.tasks_count * 5; // 5 cycles per task
-    const intervalMinutes = Math.floor((order.duration_hours * 60) / totalCycles);
-    
+    const intervalMinutes = Math.floor(
+      (order.duration_hours * 60) / totalCycles,
+    );
+
     for (let i = 0; i < order.tasks_count; i++) {
       // Generate a unique wallet for this task
       const wallet = Keypair.generate();
-      
+
       tasks.push({
         order_id: order.id,
         wallet_address: wallet.publicKey.toString(),
@@ -109,7 +128,7 @@ export class VolumeEngineService implements IVolumeEngineService {
         current_volume: 0,
         interval_minutes: Math.max(1, intervalMinutes), // At least 1 minute
         cycles_completed: 0,
-        total_cycles: 5 // 3 buys + 2 sells per task
+        total_cycles: 5, // 3 buys + 2 sells per task
       });
     }
 
@@ -128,7 +147,7 @@ export class VolumeEngineService implements IVolumeEngineService {
     const result = {
       processedTasks: 0,
       completedOrders: [] as string[],
-      errors: [] as string[]
+      errors: [] as string[],
     };
 
     try {
@@ -140,21 +159,23 @@ export class VolumeEngineService implements IVolumeEngineService {
         try {
           const orderResult = await this.processOrderTasks(order.id);
           result.processedTasks += orderResult.processedTasks;
-          
+
           if (orderResult.orderCompleted) {
             result.completedOrders.push(order.id);
           }
         } catch (error) {
-          const errorMsg = `Error processing order ${order.id}: ${error instanceof Error ? error.message : 'Unknown error'}`;
+          const errorMsg = `Error processing order ${order.id}: ${error instanceof Error ? error.message : "Unknown error"}`;
           console.error(`❌ ${errorMsg}`);
           result.errors.push(errorMsg);
         }
       }
 
-      console.log(`✅ Processed ${result.processedTasks} tasks, completed ${result.completedOrders.length} orders`);
+      console.log(
+        `✅ Processed ${result.processedTasks} tasks, completed ${result.completedOrders.length} orders`,
+      );
       return result;
     } catch (error) {
-      console.error('❌ Error in processAllPendingTasks:', error);
+      console.error("❌ Error in processAllPendingTasks:", error);
       throw error;
     }
   }
@@ -174,11 +195,11 @@ export class VolumeEngineService implements IVolumeEngineService {
     // Get current task states
     const currentTasks = await this.supabaseService.getOrderTasks(orderId);
     let processedTasks = 0;
-    
+
     for (const task of currentTasks) {
       // Skip completed tasks
       if (task.status === TaskStatus.COMPLETED) continue;
-      
+
       // Check if it's time to execute this task
       const shouldExecute = await this.shouldExecuteTask(task);
       if (!shouldExecute) continue;
@@ -187,7 +208,7 @@ export class VolumeEngineService implements IVolumeEngineService {
         // Update task status to running
         if (task.status === TaskStatus.PENDING) {
           await this.supabaseService.updateVolumeTask(task.id, {
-            status: TaskStatus.RUNNING
+            status: TaskStatus.RUNNING,
           });
         }
 
@@ -203,25 +224,31 @@ export class VolumeEngineService implements IVolumeEngineService {
           cycles_completed: newCyclesCompleted,
           current_volume: newVolume,
           last_transaction_at: new Date().toISOString(),
-          status: newCyclesCompleted >= task.total_cycles ? TaskStatus.COMPLETED : TaskStatus.RUNNING
+          status:
+            newCyclesCompleted >= task.total_cycles
+              ? TaskStatus.COMPLETED
+              : TaskStatus.RUNNING,
         });
 
         processedTasks++;
-        console.log(`📈 Executed cycle ${newCyclesCompleted}/${task.total_cycles} for task ${task.id}`);
-
+        console.log(
+          `📈 Executed cycle ${newCyclesCompleted}/${task.total_cycles} for task ${task.id}`,
+        );
       } catch (error) {
         console.error(`❌ Error executing task ${task.id}:`, error);
-        
+
         await this.supabaseService.updateVolumeTask(task.id, {
-          status: TaskStatus.FAILED
+          status: TaskStatus.FAILED,
         });
       }
     }
 
     // Check if all tasks are completed
     const updatedTasks = await this.supabaseService.getOrderTasks(orderId);
-    const completedTasks = updatedTasks.filter(task => task.status === TaskStatus.COMPLETED);
-    
+    const completedTasks = updatedTasks.filter(
+      (task) => task.status === TaskStatus.COMPLETED,
+    );
+
     if (completedTasks.length === updatedTasks.length) {
       // All tasks completed, stop the order
       await this.stopVolumeGeneration(orderId);
@@ -234,55 +261,63 @@ export class VolumeEngineService implements IVolumeEngineService {
   private async shouldExecuteTask(task: VolumeTask): Promise<boolean> {
     // If task is pending, execute immediately
     if (task.status === TaskStatus.PENDING) return true;
-    
+
     // If task is completed, don't execute
-    if (task.status === TaskStatus.COMPLETED || task.status === TaskStatus.FAILED) return false;
+    if (
+      task.status === TaskStatus.COMPLETED ||
+      task.status === TaskStatus.FAILED
+    )
+      return false;
 
     // Check if enough time has passed since last execution
     if (!task.last_transaction_at) return true;
-    
+
     const lastExecution = new Date(task.last_transaction_at);
     const now = new Date();
-    const minutesSinceLastExecution = (now.getTime() - lastExecution.getTime()) / (1000 * 60);
-    
+    const minutesSinceLastExecution =
+      (now.getTime() - lastExecution.getTime()) / (1000 * 60);
+
     return minutesSinceLastExecution >= task.interval_minutes;
   }
 
-  private async executeBuySellCycle(task: VolumeTask, order: VolumeOrder): Promise<void> {
+  private async executeBuySellCycle(
+    task: VolumeTask,
+    _order: VolumeOrder,
+  ): Promise<void> {
     // This is a simulation of buy/sell transactions
     // In production, you would implement actual Solana transactions here
-    
+
     console.log(`🔄 Simulating buy/sell cycle for task ${task.id}`);
-    
+
     // Simulate random transaction amounts
     const baseAmount = 0.1; // Base SOL amount
     const randomMultiplier = 0.5 + Math.random(); // 0.5x to 1.5x variation
     const transactionAmount = baseAmount * randomMultiplier;
-    
+
     // Simulate buy transaction
     const buyTransaction = {
       task_id: task.id,
       signature: this.generateMockSignature(),
-      type: 'buy' as const,
+      type: "buy" as const,
       amount_sol: transactionAmount,
       amount_tokens: transactionAmount * 1000, // Mock token amount
-      price: 0.001 // Mock price
+      price: 0.001, // Mock price
     };
 
     await this.supabaseService.createTransaction(buyTransaction);
-    
+
     // Wait a bit between buy and sell
-    await new Promise(resolve => setTimeout(resolve, 1000));
-    
+    await new Promise((resolve) => setTimeout(resolve, 1000));
+
     // Simulate sell transaction (smaller amount)
     const sellAmount = transactionAmount * 0.8; // Sell 80% of what we bought
     const sellTransaction = {
       task_id: task.id,
       signature: this.generateMockSignature(),
-      type: 'sell' as const,
+      type: "sell" as const,
       amount_sol: sellAmount,
       amount_tokens: sellAmount * 1000,
-      price: 0.001
+      price: 0.001,
     };
 
     await this.supabaseService.createTransaction(sellTransaction);
@@ -290,8 +325,9 @@ export class VolumeEngineService implements IVolumeEngineService {
 
   private generateMockSignature(): string {
     // Generate a mock transaction signature for simulation
-    const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-    let result = '';
+    const chars =
+      "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+    let result = "";
     for (let i = 0; i < 88; i++) {
       result += chars.charAt(Math.floor(Math.random() * chars.length));
     }
@@ -304,10 +340,12 @@ export class VolumeEngineService implements IVolumeEngineService {
   private async getRunningOrders(): Promise<VolumeOrder[]> {
     try {
       // Get all orders with running status
-      const orders = await this.supabaseService.getOrdersByStatus(OrderStatus.RUNNING);
+      const orders = await this.supabaseService.getOrdersByStatus(
+        OrderStatus.RUNNING,
+      );
       return orders;
     } catch (error) {
-      console.error('Error getting running orders:', error);
+      console.error("Error getting running orders:", error);
       return [];
     }
   }
@@ -316,8 +354,8 @@ export class VolumeEngineService implements IVolumeEngineService {
    * Initialize volume engine
    */
   async initialize(): Promise<void> {
-    console.log('🔄 Initializing Volume Engine (Stateless Mode)...');
-    console.log('📊 Volume processing will be handled by cron jobs');
-    console.log('✅ Volume Engine initialized');
+    console.log("🔄 Initializing Volume Engine (Stateless Mode)...");
+    console.log("📊 Volume processing will be handled by cron jobs");
+    console.log("✅ Volume Engine initialized");
   }
 }
