@@ -1,13 +1,13 @@
-import { Telegraf, Context, Markup } from 'telegraf';
-import { injectable, inject } from 'inversify';
-import { BaseService } from './base.service';
-import { TYPES } from '../types';
-import { ErrorService, ErrorType } from './error.service';
-import { VolumeOrderService } from './volume-order.service';
-import { SolanaService } from './solana.service';
-import { PaymentService } from './payment.service';
-import { volumeBotSettings } from '../config';
-import { UserSession, OrderStatus } from '../interfaces';
+import { Telegraf, Context, Markup } from "telegraf";
+import { injectable, inject } from "inversify";
+import { BaseService } from "./base.service";
+import { TYPES } from "../types";
+import { ErrorService, ErrorType } from "./error.service";
+import { VolumeOrderService } from "./volume-order.service";
+import { SolanaService } from "./solana.service";
+import { PaymentService } from "./payment.service";
+import { volumeBotSettings } from "../config";
+import { UserSession, OrderStatus } from "../interfaces";
 
 @injectable()
 export class PriceGoUpBotService extends BaseService {
@@ -22,7 +22,7 @@ export class PriceGoUpBotService extends BaseService {
     @inject(TYPES.ErrorService) errorService: ErrorService,
     @inject(TYPES.VolumeOrderService) volumeOrderService: VolumeOrderService,
     @inject(TYPES.SolanaService) solanaService: SolanaService,
-    @inject(TYPES.PaymentService) paymentService: PaymentService
+    @inject(TYPES.PaymentService) paymentService: PaymentService,
   ) {
     super(bot);
     this.errorService = errorService;
@@ -32,50 +32,63 @@ export class PriceGoUpBotService extends BaseService {
   }
 
   async initialize(): Promise<void> {
-    console.log('🔄 Setting up PriceGoUpBot commands...');
-    
+    console.log("🔄 Setting up PriceGoUpBot commands...");
+
     try {
       // Register commands with Telegram
       await this.bot.telegram.setMyCommands([
-        { command: 'start', description: 'Start the bot and create volume orders' },
-        { command: 'orders', description: 'View your active orders' },
-        { command: 'help', description: 'Get help with the bot' }
+        {
+          command: "start",
+          description: "Start the bot and create volume orders",
+        },
+        { command: "orders", description: "View your active orders" },
+        { command: "help", description: "Get help with the bot" },
       ]);
-      console.log('✅ Bot commands registered successfully');
+      console.log("✅ Bot commands registered successfully");
     } catch (error) {
-      console.warn('⚠️ Could not register bot commands:', error);
+      console.warn("⚠️ Could not register bot commands:", error);
     }
 
     // Setup command handlers
-    this.bot.command('start', (ctx) => this.handleStartCommand(ctx));
-    this.bot.command('orders', (ctx) => this.handleOrdersCommand(ctx));
-    this.bot.command('help', (ctx) => this.handleHelpCommand(ctx));
+    this.bot.command("start", (ctx) => this.handleStartCommand(ctx));
+    this.bot.command("orders", (ctx) => this.handleOrdersCommand(ctx));
+    this.bot.command("help", (ctx) => this.handleHelpCommand(ctx));
 
     // Setup callback query handlers
     this.bot.action(/^volume_(.+)$/, (ctx) => this.handleVolumeSelection(ctx));
-    this.bot.action(/^duration_(.+)$/, (ctx) => this.handleDurationSelection(ctx));
-    this.bot.action('separator_amount', (ctx) => this.handleSeparator(ctx));
-    this.bot.action('separator_duration', (ctx) => this.handleSeparator(ctx));
-    this.bot.action('continue', (ctx) => this.handleContinue(ctx));
-    this.bot.action('back', (ctx) => this.handleBack(ctx));
+    this.bot.action(/^duration_(.+)$/, (ctx) =>
+      this.handleDurationSelection(ctx),
+    );
+    this.bot.action("separator_amount", (ctx) => this.handleSeparator(ctx));
+    this.bot.action("separator_duration", (ctx) => this.handleSeparator(ctx));
+    this.bot.action("continue", (ctx) => this.handleContinue(ctx));
+    this.bot.action("back", (ctx) => this.handleBack(ctx));
     this.bot.action(/^pool_(.+)$/, (ctx) => this.handlePoolSelection(ctx));
-    this.bot.action('confirm_order', (ctx) => this.handleOrderConfirmation(ctx));
-    this.bot.action('check_payment', (ctx) => this.handlePaymentCheck(ctx));
-    this.bot.action('cancel_order', (ctx) => this.handleOrderCancellation(ctx));
+    this.bot.action("confirm_order", (ctx) =>
+      this.handleOrderConfirmation(ctx),
+    );
+    this.bot.action("check_payment", (ctx) => this.handlePaymentCheck(ctx));
+    this.bot.action("cancel_order", (ctx) => this.handleOrderCancellation(ctx));
 
     // Order management handlers
     this.bot.action(/^view_order_(.+)$/, (ctx) => this.handleViewOrder(ctx));
     this.bot.action(/^pause_order_(.+)$/, (ctx) => this.handlePauseOrder(ctx));
-    this.bot.action(/^resume_order_(.+)$/, (ctx) => this.handleResumeOrder(ctx));
+    this.bot.action(/^resume_order_(.+)$/, (ctx) =>
+      this.handleResumeOrder(ctx),
+    );
     this.bot.action(/^stop_order_(.+)$/, (ctx) => this.handleStopOrder(ctx));
-    this.bot.action(/^delete_order_(.+)$/, (ctx) => this.handleDeleteOrder(ctx));
-    this.bot.action(/^confirm_delete_(.+)$/, (ctx) => this.handleConfirmDelete(ctx));
+    this.bot.action(/^delete_order_(.+)$/, (ctx) =>
+      this.handleDeleteOrder(ctx),
+    );
+    this.bot.action(/^confirm_delete_(.+)$/, (ctx) =>
+      this.handleConfirmDelete(ctx),
+    );
     this.bot.action(/^orders_page_(.+)$/, (ctx) => this.handleOrdersPage(ctx));
 
     // Setup text message handler for token addresses
-    this.bot.on('text', (ctx) => this.handleTextMessage(ctx));
-    
-    console.log('✅ PriceGoUpBot commands set up successfully');
+    this.bot.on("text", (ctx) => this.handleTextMessage(ctx));
+
+    console.log("✅ PriceGoUpBot commands set up successfully");
   }
 
   /**
@@ -89,7 +102,7 @@ export class PriceGoUpBotService extends BaseService {
       // Initialize user session
       this.userSessions.set(userId, {
         userId,
-        currentStep: 'volume_selection'
+        currentStep: "volume_selection",
       });
 
       await this.showVolumeAndDurationSelection(ctx);
@@ -100,9 +113,9 @@ export class PriceGoUpBotService extends BaseService {
         ErrorType.GENERAL,
         this.errorService.createErrorContext(
           ctx.from?.id?.toString(),
-          'handleStartCommand',
-          { command: 'start' }
-        )
+          "handleStartCommand",
+          { command: "start" },
+        ),
       );
     }
   }
@@ -115,54 +128,70 @@ export class PriceGoUpBotService extends BaseService {
     if (!session) return;
 
     // Volume buttons
-    const volumeButtons = volumeBotSettings.volumePackages.map(volume => {
+    const volumeButtons = volumeBotSettings.volumePackages.map((volume) => {
       const formattedVolume = this.formatVolume(volume);
       const isPopular = volume === 1000000;
       const isTrending = volume === 5000000 || volume === 10000000;
       const isSelected = session.selectedVolume === volume;
-      
+
       let label = formattedVolume;
-      if (isPopular) label += ' ⭐ (Popular)';
-      if (isTrending) label += ' 🔥 (Trending)';
+      if (isPopular) label += " ⭐ (Popular)";
+      if (isTrending) label += " 🔥 (Trending)";
       if (isSelected) label = `✅ ${label}`;
-      
+
       return Markup.button.callback(label, `volume_${volume}`);
     });
 
     // Duration buttons
-    const durationButtons = volumeBotSettings.durations.map(hours => {
+    const durationButtons = volumeBotSettings.durations.map((hours) => {
       const isSelected = session.selectedDuration === hours;
-      const label = isSelected ? `✅ ${this.formatDuration(hours)}` : this.formatDuration(hours);
+      const label = isSelected
+        ? `✅ ${this.formatDuration(hours)}`
+        : this.formatDuration(hours);
       return Markup.button.callback(label, `duration_${hours}`);
     });
 
     // Arrange volume buttons in rows of 3
     const keyboard = [];
-    
+
     // Add volume header
-    keyboard.push([Markup.button.callback('Choose Volume:', 'separator_amount')]);
-    
+    keyboard.push([
+      Markup.button.callback("Choose Volume:", "separator_amount"),
+    ]);
+
     for (let i = 0; i < volumeButtons.length; i += 3) {
       keyboard.push(volumeButtons.slice(i, i + 3));
     }
 
     // Add duration separator
-    keyboard.push([Markup.button.callback('Set Duration:', 'separator_duration')]);
+    keyboard.push([
+      Markup.button.callback("Set Duration:", "separator_duration"),
+    ]);
 
     // Arrange duration buttons in rows of 3
     for (let i = 0; i < durationButtons.length; i += 3) {
       keyboard.push(durationButtons.slice(i, i + 3));
     }
-    
+
     // Add Continue button if both are selected
     if (session.selectedVolume && session.selectedDuration) {
-      keyboard.push([Markup.button.callback('✅ Continue', 'continue')]);
+      keyboard.push([Markup.button.callback("✅ Continue", "continue")]);
     }
 
-    const tasksCount = session.selectedVolume && session.selectedDuration ? 
-      this.calculateTasksCount(session.selectedVolume, session.selectedDuration) : 0;
-    const totalCost = session.selectedVolume && session.selectedDuration ? 
-      await this.calculateTotalCost(session.selectedVolume, session.selectedDuration) : '0';
+    const tasksCount =
+      session.selectedVolume && session.selectedDuration
+        ? this.calculateTasksCount(
+            session.selectedVolume,
+            session.selectedDuration,
+          )
+        : 0;
+    const totalCost =
+      session.selectedVolume && session.selectedDuration
+        ? await this.calculateTotalCost(
+            session.selectedVolume,
+            session.selectedDuration,
+          )
+        : "0";
 
     const message = `🚀 **PriceGoUpBot - Volume Generation**
 
@@ -177,29 +206,33 @@ Package price includes all expenses - 0% hidden fees!
 🧠 All estimates below are based on Raydium's 0.25% swap fee. Pools with higher swap fees will result in lower volume and shorter duration; our swap fee is 0%.
 
 🤖 Number of multi-tasks: ${tasksCount}
-📈 Volume Selected: ${session.selectedVolume ? this.formatVolume(session.selectedVolume) : '$0'}
-⏳ Duration Selected: ${session.selectedDuration ? this.formatDuration(session.selectedDuration) : '0'}
+📈 Volume Selected: ${session.selectedVolume ? this.formatVolume(session.selectedVolume) : "$0"}
+⏳ Duration Selected: ${session.selectedDuration ? this.formatDuration(session.selectedDuration) : "0"}
 
 💸 Total to pay: ${totalCost} SOL`;
 
     if (ctx.callbackQuery) {
       try {
         await ctx.editMessageText(message, {
-          parse_mode: 'Markdown',
-          reply_markup: Markup.inlineKeyboard(keyboard).reply_markup
+          parse_mode: "Markdown",
+          reply_markup: Markup.inlineKeyboard(keyboard).reply_markup,
         });
-      } catch (error: any) {
+      } catch (error: unknown) {
         // If message is not modified, ignore the error (content is the same)
-        if (error.code === 400 && error.description?.includes('message is not modified')) {
-          console.log('📝 Message content unchanged, skipping edit');
+        const errorObj = error as { code?: number; description?: string };
+        if (
+          errorObj.code === 400 &&
+          errorObj.description?.includes("message is not modified")
+        ) {
+          console.log("📝 Message content unchanged, skipping edit");
           return;
         }
         throw error;
       }
     } else {
       await ctx.reply(message, {
-        parse_mode: 'Markdown',
-        reply_markup: Markup.inlineKeyboard(keyboard).reply_markup
+        parse_mode: "Markdown",
+        reply_markup: Markup.inlineKeyboard(keyboard).reply_markup,
       });
     }
   }
@@ -210,8 +243,8 @@ Package price includes all expenses - 0% hidden fees!
   async handleVolumeSelection(ctx: Context): Promise<void> {
     try {
       const callbackQuery = ctx.callbackQuery;
-      if (!callbackQuery || !('data' in callbackQuery)) return;
-      
+      if (!callbackQuery || !("data" in callbackQuery)) return;
+
       const match = callbackQuery.data?.match(/^volume_(.+)$/);
       if (!match) return;
 
@@ -221,28 +254,36 @@ Package price includes all expenses - 0% hidden fees!
 
       // Optimistic update: Update UI immediately
       session.selectedVolume = volume;
-      console.log(`📊 User ${session.userId} selected volume: ${this.formatVolume(volume)}`);
-      
+      console.log(
+        `📊 User ${session.userId} selected volume: ${this.formatVolume(volume)}`,
+      );
+
       // Show updated UI instantly (optimistic)
       await this.showVolumeAndDurationSelection(ctx);
-      
+
       // Background database operations (don't wait)
-      this.handleVolumeSelectionBackground(session, ctx, volume).catch(error => {
-        console.error('❌ Background volume selection error:', error);
-      });
-      
+      this.handleVolumeSelectionBackground(session, ctx, volume).catch(
+        (error) => {
+          console.error("❌ Background volume selection error:", error);
+        },
+      );
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
     }
   }
 
-  private async handleVolumeSelectionBackground(session: any, ctx: Context, volume: number): Promise<void> {
-    console.log(`📊 Current session state: volume=${session.selectedVolume}, duration=${session.selectedDuration}`);
-    
+  private async handleVolumeSelectionBackground(
+    session: UserSession,
+    ctx: Context,
+    _volume: number,
+  ): Promise<void> {
+    console.log(
+      `📊 Current session state: volume=${session.selectedVolume}, duration=${session.selectedDuration}`,
+    );
+
     // Create draft order if both volume and duration are selected
     await this.createDraftOrderIfReady(session, ctx);
   }
-
 
   /**
    * Handle duration selection
@@ -250,8 +291,8 @@ Package price includes all expenses - 0% hidden fees!
   async handleDurationSelection(ctx: Context): Promise<void> {
     try {
       const callbackQuery = ctx.callbackQuery;
-      if (!callbackQuery || !('data' in callbackQuery)) return;
-      
+      if (!callbackQuery || !("data" in callbackQuery)) return;
+
       const match = callbackQuery.data?.match(/^duration_(.+)$/);
       if (!match) return;
 
@@ -261,24 +302,33 @@ Package price includes all expenses - 0% hidden fees!
 
       // Optimistic update: Update UI immediately
       session.selectedDuration = duration;
-      console.log(`⏰ User ${session.userId} selected duration: ${this.formatDuration(duration)}`);
-      
+      console.log(
+        `⏰ User ${session.userId} selected duration: ${this.formatDuration(duration)}`,
+      );
+
       // Show updated UI instantly (optimistic)
       await this.showVolumeAndDurationSelection(ctx);
-      
+
       // Background database operations (don't wait)
-      this.handleDurationSelectionBackground(session, ctx, duration).catch(error => {
-        console.error('❌ Background duration selection error:', error);
-      });
-      
+      this.handleDurationSelectionBackground(session, ctx, duration).catch(
+        (error) => {
+          console.error("❌ Background duration selection error:", error);
+        },
+      );
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
     }
   }
 
-  private async handleDurationSelectionBackground(session: any, ctx: Context, duration: number): Promise<void> {
-    console.log(`📊 Current session state: volume=${session.selectedVolume}, duration=${session.selectedDuration}`);
-    
+  private async handleDurationSelectionBackground(
+    session: UserSession,
+    ctx: Context,
+    _duration: number,
+  ): Promise<void> {
+    console.log(
+      `📊 Current session state: volume=${session.selectedVolume}, duration=${session.selectedDuration}`,
+    );
+
     // Create draft order if both volume and duration are selected
     await this.createDraftOrderIfReady(session, ctx);
   }
@@ -293,25 +343,36 @@ Package price includes all expenses - 0% hidden fees!
   /**
    * Create draft order if both volume and duration are selected
    */
-  private async createDraftOrderIfReady(session: any, ctx: Context): Promise<void> {
-    if (session.selectedVolume && session.selectedDuration && !session.orderData) {
+  private async createDraftOrderIfReady(
+    session: UserSession,
+    ctx: Context,
+  ): Promise<void> {
+    if (
+      session.selectedVolume &&
+      session.selectedDuration &&
+      !session.orderData
+    ) {
       try {
-        console.log(`🔄 Creating draft order for user ${session.userId}: ${session.selectedVolume} volume, ${session.selectedDuration}h duration`);
-        
+        console.log(
+          `🔄 Creating draft order for user ${session.userId}: ${session.selectedVolume} volume, ${session.selectedDuration}h duration`,
+        );
+
         const draftOrder = await this.volumeOrderService.createOrder({
           user_id: session.userId,
           username: ctx.from?.username,
-          token_address: 'PENDING', // Placeholder until user provides
-          pool_address: 'PENDING',  // Placeholder until user selects
+          token_address: "PENDING", // Placeholder until user provides
+          pool_address: "PENDING", // Placeholder until user selects
           volume_target: session.selectedVolume,
           duration_hours: session.selectedDuration,
-          status: OrderStatus.PENDING_PAYMENT
+          status: OrderStatus.PENDING_PAYMENT,
         });
-        
+
         session.orderData = draftOrder;
-        console.log(`✅ Created draft order ${draftOrder.id} for user ${session.userId}`);
+        console.log(
+          `✅ Created draft order ${draftOrder.id} for user ${session.userId}`,
+        );
       } catch (error) {
-        console.error('❌ Error creating draft order:', error);
+        console.error("❌ Error creating draft order:", error);
       }
     }
   }
@@ -323,18 +384,22 @@ Package price includes all expenses - 0% hidden fees!
     try {
       const session = await this.getUserSession(ctx.from?.id?.toString());
       if (!session || !session.selectedVolume || !session.selectedDuration) {
-        await ctx.answerCbQuery('Please select both volume and duration first!');
+        await ctx.answerCbQuery(
+          "Please select both volume and duration first!",
+        );
         return;
       }
 
       // Order should already be created by createDraftOrderIfReady
       if (!session.orderData) {
-        await ctx.answerCbQuery('Order data missing. Please try again.');
+        await ctx.answerCbQuery("Order data missing. Please try again.");
         return;
       }
 
-      session.currentStep = 'token_address';
-      console.log(`🔄 Continue clicked - setting currentStep to 'token_address' for user ${session.userId}`);
+      session.currentStep = "token_address";
+      console.log(
+        `🔄 Continue clicked - setting currentStep to 'token_address' for user ${session.userId}`,
+      );
       await this.showOrderSummary(ctx);
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
@@ -346,17 +411,26 @@ Package price includes all expenses - 0% hidden fees!
    */
   async showOrderSummary(ctx: Context): Promise<void> {
     const session = await this.getUserSession(ctx.from?.id?.toString());
-    if (!session) return;
+    if (!session || !session.selectedVolume || !session.selectedDuration)
+      return;
 
-    const tasksCount = this.calculateTasksCount(session.selectedVolume!, session.selectedDuration!);
-    const costData = await this.paymentService.calculateOrderCost(session.selectedVolume!, session.selectedDuration!);
-    const intervalMinutes = Math.floor(session.selectedDuration! * 60 / tasksCount);
+    const tasksCount = this.calculateTasksCount(
+      session.selectedVolume,
+      session.selectedDuration,
+    );
+    const costData = await this.paymentService.calculateOrderCost(
+      session.selectedVolume,
+      session.selectedDuration,
+    );
+    const intervalMinutes = Math.floor(
+      (session.selectedDuration * 60) / tasksCount,
+    );
 
     const message = `Your configuration summary:
 
 Each task will repeatedly buy and sell tokens with specific intervals between transactions.
 
-⚡ ${this.formatVolume(session.selectedVolume!)} - ${this.formatDuration(session.selectedDuration!)}
+⚡ ${this.formatVolume(session.selectedVolume)} - ${this.formatDuration(session.selectedDuration)}
 🤖 Tasks - Interval
 ${tasksCount}x ${costData.costPerTask} SOL - ⏳ ${this.formatInterval(intervalMinutes)}
 
@@ -369,10 +443,10 @@ Total: 🟪 ${costData.totalCost} SOL
 🔽 Please send as a chat message.`;
 
     await ctx.editMessageText(message, {
-      parse_mode: 'Markdown',
+      parse_mode: "Markdown",
       reply_markup: Markup.inlineKeyboard([
-        [Markup.button.callback('🔙 Back', 'back')]
-      ]).reply_markup
+        [Markup.button.callback("🔙 Back", "back")],
+      ]).reply_markup,
     });
   }
 
@@ -382,12 +456,16 @@ Total: 🟪 ${costData.totalCost} SOL
   async handleTextMessage(ctx: Context): Promise<void> {
     try {
       const session = await this.getUserSession(ctx.from?.id?.toString());
-      const text = (ctx as any).message?.text?.trim();
-      
+      const text = (
+        ctx as { message?: { text?: string } }
+      ).message?.text?.trim();
+
       console.log(`📝 Received text message: "${text}"`);
-      console.log(`📊 Session state: step=${session?.currentStep}, user=${session?.userId}`);
-      
-      if (!session || session.currentStep !== 'token_address') {
+      console.log(
+        `📊 Session state: step=${session?.currentStep}, user=${session?.userId}`,
+      );
+
+      if (!session || session.currentStep !== "token_address") {
         console.log(`❌ Ignoring text message - wrong step or no session`);
         return;
       }
@@ -397,36 +475,46 @@ Total: 🟪 ${costData.totalCost} SOL
       // Validate token address
       const isValid = await this.solanaService.validateTokenAddress(text);
       if (!isValid) {
-        await ctx.reply('❌ Invalid token address. Please send a valid Solana token contract address.');
+        await ctx.reply(
+          "❌ Invalid token address. Please send a valid Solana token contract address.",
+        );
         return;
       }
 
       // Optimistic update: Update session and show loading immediately
       session.tokenAddress = text;
-      session.currentStep = 'pool_selection';
-      
+      session.currentStep = "pool_selection";
+
       // Show "searching for pools" message immediately
-      await ctx.reply('🔍 Validating token and searching for liquidity pools...');
-      
+      await ctx.reply(
+        "🔍 Validating token and searching for liquidity pools...",
+      );
+
       // Background operations (validation, DB update, pool fetching)
-      this.handleTokenAddressBackground(session, ctx, text).catch(error => {
-        console.error('❌ Background token address processing error:', error);
-        ctx.reply('❌ Error processing token address. Please try again.');
+      this.handleTokenAddressBackground(session, ctx, text).catch((error) => {
+        console.error("❌ Background token address processing error:", error);
+        ctx.reply("❌ Error processing token address. Please try again.");
       });
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
     }
   }
 
-  private async handleTokenAddressBackground(session: any, ctx: Context, text: string): Promise<void> {
+  private async handleTokenAddressBackground(
+    session: UserSession,
+    ctx: Context,
+    text: string,
+  ): Promise<void> {
     // Update the draft order with the token address
     if (session.orderData && session.orderData.id) {
       await this.volumeOrderService.updateOrder(session.orderData.id, {
-        token_address: text
+        token_address: text,
       });
-      console.log(`📝 Updated order ${session.orderData.id} with token address`);
+      console.log(
+        `📝 Updated order ${session.orderData.id} with token address`,
+      );
     }
-    
+
     // Show pool selection
     await this.showPoolSelection(ctx);
   }
@@ -439,10 +527,14 @@ Total: 🟪 ${costData.totalCost} SOL
     if (!session || !session.tokenAddress) return;
 
     try {
-      const pools = await this.solanaService.getLiquidityPools(session.tokenAddress);
-      
+      const pools = await this.solanaService.getLiquidityPools(
+        session.tokenAddress,
+      );
+
       if (pools.length === 0) {
-        await ctx.reply('❌ No liquidity pools found for this token. Please check the token address and try again.');
+        await ctx.reply(
+          "❌ No liquidity pools found for this token. Please check the token address and try again.",
+        );
         return;
       }
 
@@ -450,15 +542,17 @@ Total: 🟪 ${costData.totalCost} SOL
         return Markup.button.callback(`-${index + 1}-`, `pool_${pool.address}`);
       });
 
-      const keyboard = poolButtons.map(button => [button]);
-      keyboard.push([Markup.button.callback('🔙 Back', 'back')]);
+      const keyboard = poolButtons.map((button) => [button]);
+      keyboard.push([Markup.button.callback("🔙 Back", "back")]);
 
-      let message = '💧 Select the liquidity pool you want to use:\n\n';
-      
+      let message = "💧 Select the liquidity pool you want to use:\n\n";
+
       pools.slice(0, 5).forEach((pool, index) => {
         console.log(`🔍 Pool ${index + 1} address: ${pool.address}`);
-        console.log(`🔗 Dexscreener link: https://dexscreener.com/solana/${pool.address}`);
-        
+        console.log(
+          `🔗 Dexscreener link: https://dexscreener.com/solana/${pool.address}`,
+        );
+
         message += `-${index + 1}-\n`;
         message += `💧 LP Type: ${pool.pool_type.toUpperCase()}\n`;
         message += `📄 Pool address: ${pool.address}\n`;
@@ -470,8 +564,8 @@ Total: 🟪 ${costData.totalCost} SOL
       });
 
       await ctx.reply(message, {
-        parse_mode: 'Markdown',
-        reply_markup: Markup.inlineKeyboard(keyboard).reply_markup
+        parse_mode: "Markdown",
+        reply_markup: Markup.inlineKeyboard(keyboard).reply_markup,
       });
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
@@ -484,8 +578,8 @@ Total: 🟪 ${costData.totalCost} SOL
   async handlePoolSelection(ctx: Context): Promise<void> {
     try {
       const callbackQuery = ctx.callbackQuery;
-      if (!callbackQuery || !('data' in callbackQuery)) return;
-      
+      if (!callbackQuery || !("data" in callbackQuery)) return;
+
       const match = callbackQuery.data?.match(/^pool_(.+)$/);
       if (!match) return;
 
@@ -495,25 +589,30 @@ Total: 🟪 ${costData.totalCost} SOL
 
       // Optimistic update: Update UI immediately
       session.selectedPool = poolAddress;
-      session.currentStep = 'order_review';
-      
+      session.currentStep = "order_review";
+
       // Show order review immediately
       await this.showOrderReview(ctx);
-      
+
       // Background database update
-      this.handlePoolSelectionBackground(session, poolAddress).catch(error => {
-        console.error('❌ Background pool selection error:', error);
-      });
+      this.handlePoolSelectionBackground(session, poolAddress).catch(
+        (error) => {
+          console.error("❌ Background pool selection error:", error);
+        },
+      );
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
     }
   }
 
-  private async handlePoolSelectionBackground(session: any, poolAddress: string): Promise<void> {
+  private async handlePoolSelectionBackground(
+    session: UserSession,
+    poolAddress: string,
+  ): Promise<void> {
     // Update the draft order with the pool address
     if (session.orderData && session.orderData.id) {
       await this.volumeOrderService.updateOrder(session.orderData.id, {
-        pool_address: poolAddress
+        pool_address: poolAddress,
       });
       console.log(`📝 Updated order ${session.orderData.id} with pool address`);
     }
@@ -524,10 +623,22 @@ Total: 🟪 ${costData.totalCost} SOL
    */
   async showOrderReview(ctx: Context): Promise<void> {
     const session = await this.getUserSession(ctx.from?.id?.toString());
-    if (!session || !session.selectedVolume || !session.selectedDuration || !session.tokenAddress || !session.selectedPool) return;
+    if (
+      !session ||
+      !session.selectedVolume ||
+      !session.selectedDuration ||
+      !session.tokenAddress ||
+      !session.selectedPool
+    )
+      return;
 
-    const costData = await this.paymentService.calculateOrderCost(session.selectedVolume, session.selectedDuration);
-    const intervalMinutes = Math.floor(session.selectedDuration * 60 / costData.tasksCount);
+    const costData = await this.paymentService.calculateOrderCost(
+      session.selectedVolume,
+      session.selectedDuration,
+    );
+    const intervalMinutes = Math.floor(
+      (session.selectedDuration * 60) / costData.tasksCount,
+    );
 
     const message = `📋 Review your order summary:
 
@@ -550,9 +661,9 @@ ${session.selectedPool}
 
     await ctx.reply(message, {
       reply_markup: Markup.inlineKeyboard([
-        [Markup.button.callback('💳 Pay Order', 'confirm_order')],
-        [Markup.button.callback('🔙 Back', 'back')]
-      ]).reply_markup
+        [Markup.button.callback("💳 Pay Order", "confirm_order")],
+        [Markup.button.callback("🔙 Back", "back")],
+      ]).reply_markup,
     });
   }
 
@@ -566,15 +677,18 @@ ${session.selectedPool}
 
       // Order already exists as draft, just generate payment QR code
       const order = session.orderData;
-      
+
       if (!order.payment_address || !order.total_cost) {
-        await ctx.answerCbQuery('Order data incomplete. Please start over.');
+        await ctx.answerCbQuery("Order data incomplete. Please start over.");
         return;
       }
-      
-      const qrCode = await this.paymentService.generateQRCode(order.payment_address, order.total_cost);
 
-      session.currentStep = 'payment_pending';
+      const qrCode = await this.paymentService.generateQRCode(
+        order.payment_address,
+        order.total_cost,
+      );
+
+      session.currentStep = "payment_pending";
 
       const message = `Scan the QR code above with your mobile wallet to make your payment quickly.
 
@@ -586,14 +700,16 @@ ${order.payment_address}
 
 🔽 If you've already made payment, click "Check & Continue" button below to proceed.`;
 
-      await ctx.replyWithPhoto({ source: Buffer.from(qrCode, 'base64') }, {
-        caption: message,
-        reply_markup: Markup.inlineKeyboard([
-          [Markup.button.callback('✅ Check & Continue', 'check_payment')],
-          [Markup.button.callback('❌ Cancel', 'cancel_order')]
-        ]).reply_markup
-      });
-
+      await ctx.replyWithPhoto(
+        { source: Buffer.from(qrCode, "base64") },
+        {
+          caption: message,
+          reply_markup: Markup.inlineKeyboard([
+            [Markup.button.callback("✅ Check & Continue", "check_payment")],
+            [Markup.button.callback("❌ Cancel", "cancel_order")],
+          ]).reply_markup,
+        },
+      );
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
     }
@@ -604,27 +720,44 @@ ${order.payment_address}
    */
   async handlePaymentCheck(ctx: Context): Promise<void> {
     try {
-      try { await ctx.answerCbQuery('Checking payment...'); } catch {}
+      try {
+        await ctx.answerCbQuery("Checking payment...");
+      } catch {
+        /* Ignore callback query errors */
+      }
       const session = await this.getUserSession(ctx.from?.id?.toString());
       if (!session || !session.orderData) return;
 
       const orderData = session.orderData;
-      if (!orderData.payment_address || !orderData.total_cost || !orderData.id) {
-        await ctx.answerCbQuery('Order data incomplete. Please start over.');
+      if (
+        !orderData.payment_address ||
+        !orderData.total_cost ||
+        !orderData.id
+      ) {
+        await ctx.answerCbQuery("Order data incomplete. Please start over.");
         return;
       }
 
-      console.log(`🔎 handlePaymentCheck: orderId=${orderData.id} address=${orderData.payment_address} expected=${orderData.total_cost}`);
+      console.log(
+        `🔎 handlePaymentCheck: orderId=${orderData.id} address=${orderData.payment_address} expected=${orderData.total_cost}`,
+      );
       const signature = await this.solanaService.verifyPayment(
         orderData.payment_address,
-        orderData.total_cost
+        orderData.total_cost,
       );
 
       if (signature) {
         // Acknowledge the button to stop the loading spinner
-        try { await ctx.answerCbQuery('Payment confirmed!'); } catch {}
-        await this.volumeOrderService.updateOrderStatus(orderData.id, OrderStatus.PAYMENT_CONFIRMED);
-        
+        try {
+          await ctx.answerCbQuery("Payment confirmed!");
+        } catch {
+          /* Ignore callback query errors */
+        }
+        await this.volumeOrderService.updateOrderStatus(
+          orderData.id,
+          OrderStatus.PAYMENT_CONFIRMED,
+        );
+
         await ctx.reply(`🎉 Payment confirmed! Your volume generation order is now active.
 
 Order ID: ${orderData.id}
@@ -635,11 +768,19 @@ Your bots will start generating volume shortly. Use /orders to check your order 
         // Clear session
         this.userSessions.delete(session.userId);
       } else {
-        try { await ctx.answerCbQuery('Payment not found yet. Please wait a moment and try again.'); } catch {}
-        await ctx.reply('⌛ Still waiting for payment to arrive. If you just sent it, give it ~15–30s and press "Check & Continue" again.');
+        try {
+          await ctx.answerCbQuery(
+            "Payment not found yet. Please wait a moment and try again.",
+          );
+        } catch {
+          /* Ignore callback query errors */
+        }
+        await ctx.reply(
+          '⌛ Still waiting for payment to arrive. If you just sent it, give it ~15–30s and press "Check & Continue" again.',
+        );
       }
     } catch (error) {
-      console.error('❌ handlePaymentCheck error:', error);
+      console.error("❌ handlePaymentCheck error:", error);
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
     }
   }
@@ -652,12 +793,12 @@ Your bots will start generating volume shortly. Use /orders to check your order 
     if (!session) return;
 
     switch (session.currentStep) {
-      case 'token_address':
-        session.currentStep = 'volume_selection';
+      case "token_address":
+        session.currentStep = "volume_selection";
         await this.showVolumeAndDurationSelection(ctx);
         break;
-      case 'pool_selection':
-        session.currentStep = 'token_address';
+      case "pool_selection":
+        session.currentStep = "token_address";
         await this.showOrderSummary(ctx);
         break;
       default:
@@ -687,9 +828,11 @@ Your bots will start generating volume shortly. Use /orders to check your order 
     if (!userId) return;
 
     const orders = await this.volumeOrderService.getUserOrders(userId);
-    
+
     if (orders.length === 0) {
-      await ctx.reply('You have no orders yet. Use /start to create your first volume order!');
+      await ctx.reply(
+        "You have no orders yet. Use /start to create your first volume order!",
+      );
       return;
     }
 
@@ -699,35 +842,45 @@ Your bots will start generating volume shortly. Use /orders to check your order 
     const pageOrders = orders.slice(startIdx, startIdx + pageSize);
 
     let message = `📊 Your Volume Orders (Page ${page + 1}/${totalPages}):\n\n`;
-    
+
     const keyboard = [];
-    
+
     for (const order of pageOrders) {
       const shortId = order.id.substring(0, 8);
       const statusIcon = this.getStatusIcon(order.status);
-      
+
       message += `${statusIcon} **Order ${shortId}**\n`;
       message += `📈 ${this.formatVolume(order.volume_target)} • ${this.formatDuration(order.duration_hours)}\n`;
-      message += `💸 ${order.total_cost} SOL • ${order.status.replace('_', ' ').toUpperCase()}\n`;
+      message += `💸 ${order.total_cost} SOL • ${order.status.replace("_", " ").toUpperCase()}\n`;
       message += `📅 ${new Date(order.created_at).toLocaleDateString()}\n\n`;
-      
+
       // Add control buttons for this order
       const orderButtons = [
-        Markup.button.callback('👁️ View', `view_order_${order.id}`)
+        Markup.button.callback("👁️ View", `view_order_${order.id}`),
       ];
-      
+
       if (order.status === OrderStatus.RUNNING) {
-        orderButtons.push(Markup.button.callback('⏸️ Pause', `pause_order_${order.id}`));
-        orderButtons.push(Markup.button.callback('🛑 Stop', `stop_order_${order.id}`));
+        orderButtons.push(
+          Markup.button.callback("⏸️ Pause", `pause_order_${order.id}`),
+        );
+        orderButtons.push(
+          Markup.button.callback("🛑 Stop", `stop_order_${order.id}`),
+        );
       } else if (order.status === OrderStatus.PAUSED) {
-        orderButtons.push(Markup.button.callback('▶️ Resume', `resume_order_${order.id}`));
-        orderButtons.push(Markup.button.callback('🛑 Stop', `stop_order_${order.id}`));
+        orderButtons.push(
+          Markup.button.callback("▶️ Resume", `resume_order_${order.id}`),
+        );
+        orderButtons.push(
+          Markup.button.callback("🛑 Stop", `stop_order_${order.id}`),
+        );
       } else if (order.status === OrderStatus.PAYMENT_CONFIRMED) {
         // Will be started by volume processor
       } else if (order.status === OrderStatus.PENDING_PAYMENT) {
-        orderButtons.push(Markup.button.callback('🗑️ Delete', `delete_order_${order.id}`));
+        orderButtons.push(
+          Markup.button.callback("🗑️ Delete", `delete_order_${order.id}`),
+        );
       }
-      
+
       if (orderButtons.length > 1) {
         keyboard.push(orderButtons);
       } else {
@@ -738,50 +891,68 @@ Your bots will start generating volume shortly. Use /orders to check your order 
     // Pagination controls
     const navButtons = [];
     if (page > 0) {
-      navButtons.push(Markup.button.callback('⬅️ Prev', `orders_page_${page - 1}`));
+      navButtons.push(
+        Markup.button.callback("⬅️ Prev", `orders_page_${page - 1}`),
+      );
     }
     if (page < totalPages - 1) {
-      navButtons.push(Markup.button.callback('➡️ Next', `orders_page_${page + 1}`));
+      navButtons.push(
+        Markup.button.callback("➡️ Next", `orders_page_${page + 1}`),
+      );
     }
     if (navButtons.length > 0) {
       keyboard.push(navButtons);
     }
 
     // Add refresh button
-    keyboard.push([Markup.button.callback('🔄 Refresh', `orders_page_${page}`)]);
+    keyboard.push([
+      Markup.button.callback("🔄 Refresh", `orders_page_${page}`),
+    ]);
 
     if (ctx.callbackQuery) {
       try {
         await ctx.editMessageText(message, {
-          parse_mode: 'Markdown',
-          reply_markup: Markup.inlineKeyboard(keyboard).reply_markup
+          parse_mode: "Markdown",
+          reply_markup: Markup.inlineKeyboard(keyboard).reply_markup,
         });
-      } catch (error: any) {
+      } catch (error: unknown) {
         // If message is not modified, ignore the error (content is the same)
-        if (error.code === 400 && error.description?.includes('message is not modified')) {
-          await ctx.answerCbQuery('Orders refreshed');
+        const errorObj = error as { code?: number; description?: string };
+        if (
+          errorObj.code === 400 &&
+          errorObj.description?.includes("message is not modified")
+        ) {
+          await ctx.answerCbQuery("Orders refreshed");
           return;
         }
         throw error;
       }
     } else {
       await ctx.reply(message, {
-        parse_mode: 'Markdown',
-        reply_markup: Markup.inlineKeyboard(keyboard).reply_markup
+        parse_mode: "Markdown",
+        reply_markup: Markup.inlineKeyboard(keyboard).reply_markup,
       });
     }
   }
 
   private getStatusIcon(status: string): string {
     switch (status) {
-      case OrderStatus.PENDING_PAYMENT: return '💰';
-      case OrderStatus.PAYMENT_CONFIRMED: return '✅';
-      case OrderStatus.RUNNING: return '🚀';
-      case OrderStatus.PAUSED: return '⏸️';
-      case OrderStatus.COMPLETED: return '✨';
-      case OrderStatus.CANCELLED: return '❌';
-      case OrderStatus.FAILED: return '💥';
-      default: return '❓';
+      case OrderStatus.PENDING_PAYMENT:
+        return "💰";
+      case OrderStatus.PAYMENT_CONFIRMED:
+        return "✅";
+      case OrderStatus.RUNNING:
+        return "🚀";
+      case OrderStatus.PAUSED:
+        return "⏸️";
+      case OrderStatus.COMPLETED:
+        return "✨";
+      case OrderStatus.CANCELLED:
+        return "❌";
+      case OrderStatus.FAILED:
+        return "💥";
+      default:
+        return "❓";
     }
   }
 
@@ -816,24 +987,26 @@ PriceGoUpBot is a volume generation service for Solana SPL tokens. We create org
 **Support:**
 Contact @support for help with your orders.`;
 
-    await ctx.reply(message, { parse_mode: 'Markdown' });
+    await ctx.reply(message, { parse_mode: "Markdown" });
   }
 
   // Helper methods
-  private async getUserSession(userId: string | undefined): Promise<UserSession | null> {
+  private async getUserSession(
+    userId: string | undefined,
+  ): Promise<UserSession | null> {
     if (!userId) return null;
-    
+
     // First try to get from memory
     let session = this.userSessions.get(userId);
-    
+
     // If not in memory, try to restore from database
     if (!session) {
       const recentOrders = await this.volumeOrderService.getUserOrders(userId);
       // Restore any latest pending_payment order (even if token/pool already set)
-      const pendingOrder = recentOrders.find(order => 
-        order.status === OrderStatus.PENDING_PAYMENT
+      const pendingOrder = recentOrders.find(
+        (order) => order.status === OrderStatus.PENDING_PAYMENT,
       );
-      
+
       if (pendingOrder) {
         // Restore session from database
         session = {
@@ -841,23 +1014,33 @@ Contact @support for help with your orders.`;
           currentStep: this.determineCurrentStep(pendingOrder),
           selectedVolume: pendingOrder.volume_target,
           selectedDuration: pendingOrder.duration_hours,
-          tokenAddress: pendingOrder.token_address !== 'PENDING' ? pendingOrder.token_address : undefined,
-          selectedPool: pendingOrder.pool_address !== 'PENDING' ? pendingOrder.pool_address : undefined,
-          orderData: pendingOrder
+          tokenAddress:
+            pendingOrder.token_address !== "PENDING"
+              ? pendingOrder.token_address
+              : undefined,
+          selectedPool:
+            pendingOrder.pool_address !== "PENDING"
+              ? pendingOrder.pool_address
+              : undefined,
+          orderData: pendingOrder,
         };
-        
+
         this.userSessions.set(userId, session);
         console.log(`🔄 Restored session for user ${userId} from database`);
       }
     }
-    
+
     return session || null;
   }
 
-  private determineCurrentStep(order: any): string {
-    if (order.token_address === 'PENDING') return 'token_address';
-    if (order.pool_address === 'PENDING') return 'pool_selection';
-    return 'order_review';
+  private determineCurrentStep(order: {
+    token_address?: string;
+    pool_address?: string;
+    payment_address?: string;
+  }): string {
+    if (order.token_address === "PENDING") return "token_address";
+    if (order.pool_address === "PENDING") return "pool_selection";
+    return "order_review";
   }
 
   private formatVolume(volume: number): string {
@@ -878,9 +1061,9 @@ Contact @support for help with your orders.`;
     const hours = Math.floor(minutes / 60);
     const mins = minutes % 60;
     if (hours > 0) {
-      return `${hours.toString().padStart(2, '0')}:${mins.toString().padStart(2, '0')}`;
+      return `${hours.toString().padStart(2, "0")}:${mins.toString().padStart(2, "0")}`;
     }
-    return `00:${mins.toString().padStart(2, '0')}`;
+    return `00:${mins.toString().padStart(2, "0")}`;
   }
 
   private calculateTasksCount(volume: number, duration: number): number {
@@ -888,16 +1071,25 @@ Contact @support for help with your orders.`;
     const baseTasksPerMillion = 12;
     const volumeInMillions = volume / 1000000;
     const durationFactor = 24 / duration; // Normalize to 24 hours
-    
-    return Math.max(1, Math.floor(baseTasksPerMillion * volumeInMillions * durationFactor));
+
+    return Math.max(
+      1,
+      Math.floor(baseTasksPerMillion * volumeInMillions * durationFactor),
+    );
   }
 
-  private async calculateTotalCost(volume: number, duration: number): Promise<string> {
+  private async calculateTotalCost(
+    volume: number,
+    duration: number,
+  ): Promise<string> {
     try {
-      const costData = await this.paymentService.calculateOrderCost(volume, duration);
+      const costData = await this.paymentService.calculateOrderCost(
+        volume,
+        duration,
+      );
       return costData.totalCost.toFixed(2);
     } catch {
-      return '0.00';
+      return "0.00";
     }
   }
 
@@ -906,7 +1098,7 @@ Contact @support for help with your orders.`;
     if (session) {
       this.userSessions.delete(session.userId);
     }
-    await ctx.reply('❌ Order cancelled.');
+    await ctx.reply("❌ Order cancelled.");
   }
 
   /**
@@ -915,22 +1107,22 @@ Contact @support for help with your orders.`;
   async handleViewOrder(ctx: Context): Promise<void> {
     try {
       const callbackQuery = ctx.callbackQuery;
-      if (!callbackQuery || !('data' in callbackQuery)) return;
-      
+      if (!callbackQuery || !("data" in callbackQuery)) return;
+
       const match = callbackQuery.data?.match(/^view_order_(.+)$/);
       if (!match) return;
 
       const orderId = match[1];
       const order = await this.volumeOrderService.getOrder(orderId);
-      
+
       if (!order) {
-        await ctx.answerCbQuery('Order not found');
+        await ctx.answerCbQuery("Order not found");
         return;
       }
 
       // Get task progress via volume order service
       const progress = await this.volumeOrderService.getOrderProgress(orderId);
-      
+
       const message = `📋 **Order Details**
 
 🆔 **ID:** ${order.id.substring(0, 16)}...
@@ -940,7 +1132,7 @@ Contact @support for help with your orders.`;
 ⏳ **Duration:** ${this.formatDuration(order.duration_hours)}
 🤖 **Tasks:** ${order.tasks_count}
 💸 **Cost:** ${order.total_cost} SOL
-📊 **Status:** ${order.status.replace('_', ' ').toUpperCase()}
+📊 **Status:** ${order.status.replace("_", " ").toUpperCase()}
 
 **Progress:**
 📈 Volume Generated: $${progress.totalVolume.toLocaleString()}
@@ -948,33 +1140,35 @@ Contact @support for help with your orders.`;
 🏃 Running Tasks: ${progress.runningTasks}
 
 📅 **Created:** ${new Date(order.created_at).toLocaleString()}
-${order.started_at ? `🚀 **Started:** ${new Date(order.started_at).toLocaleString()}` : ''}
-${order.completed_at ? `✨ **Completed:** ${new Date(order.completed_at).toLocaleString()}` : ''}`;
+${order.started_at ? `🚀 **Started:** ${new Date(order.started_at).toLocaleString()}` : ""}
+${order.completed_at ? `✨ **Completed:** ${new Date(order.completed_at).toLocaleString()}` : ""}`;
 
       const keyboard = [];
-      
+
       // Add action buttons based on status
       if (order.status === OrderStatus.RUNNING) {
         keyboard.push([
-          Markup.button.callback('⏸️ Pause', `pause_order_${order.id}`),
-          Markup.button.callback('🛑 Stop', `stop_order_${order.id}`)
+          Markup.button.callback("⏸️ Pause", `pause_order_${order.id}`),
+          Markup.button.callback("🛑 Stop", `stop_order_${order.id}`),
         ]);
       } else if (order.status === OrderStatus.PAUSED) {
         keyboard.push([
-          Markup.button.callback('▶️ Resume', `resume_order_${order.id}`),
-          Markup.button.callback('🛑 Stop', `stop_order_${order.id}`)
+          Markup.button.callback("▶️ Resume", `resume_order_${order.id}`),
+          Markup.button.callback("🛑 Stop", `stop_order_${order.id}`),
         ]);
       } else if (order.status === OrderStatus.PENDING_PAYMENT) {
         keyboard.push([
-          Markup.button.callback('🗑️ Delete', `delete_order_${order.id}`)
+          Markup.button.callback("🗑️ Delete", `delete_order_${order.id}`),
         ]);
       }
-      
-      keyboard.push([Markup.button.callback('🔙 Back to Orders', 'orders_page_0')]);
+
+      keyboard.push([
+        Markup.button.callback("🔙 Back to Orders", "orders_page_0"),
+      ]);
 
       await ctx.editMessageText(message, {
-        parse_mode: 'Markdown',
-        reply_markup: Markup.inlineKeyboard(keyboard).reply_markup
+        parse_mode: "Markdown",
+        reply_markup: Markup.inlineKeyboard(keyboard).reply_markup,
       });
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
@@ -987,8 +1181,8 @@ ${order.completed_at ? `✨ **Completed:** ${new Date(order.completed_at).toLoca
   async handleOrdersPage(ctx: Context): Promise<void> {
     try {
       const callbackQuery = ctx.callbackQuery;
-      if (!callbackQuery || !('data' in callbackQuery)) return;
-      
+      if (!callbackQuery || !("data" in callbackQuery)) return;
+
       const match = callbackQuery.data?.match(/^orders_page_(.+)$/);
       if (!match) return;
 
@@ -1005,15 +1199,18 @@ ${order.completed_at ? `✨ **Completed:** ${new Date(order.completed_at).toLoca
   async handlePauseOrder(ctx: Context): Promise<void> {
     try {
       const callbackQuery = ctx.callbackQuery;
-      if (!callbackQuery || !('data' in callbackQuery)) return;
-      
+      if (!callbackQuery || !("data" in callbackQuery)) return;
+
       const match = callbackQuery.data?.match(/^pause_order_(.+)$/);
       if (!match) return;
 
       const orderId = match[1];
-      await this.volumeOrderService.updateOrderStatus(orderId, 'paused' as OrderStatus);
-      
-      await ctx.answerCbQuery('Order paused');
+      await this.volumeOrderService.updateOrderStatus(
+        orderId,
+        "paused" as OrderStatus,
+      );
+
+      await ctx.answerCbQuery("Order paused");
       await this.handleViewOrder(ctx); // Refresh the view
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
@@ -1026,15 +1223,18 @@ ${order.completed_at ? `✨ **Completed:** ${new Date(order.completed_at).toLoca
   async handleResumeOrder(ctx: Context): Promise<void> {
     try {
       const callbackQuery = ctx.callbackQuery;
-      if (!callbackQuery || !('data' in callbackQuery)) return;
-      
+      if (!callbackQuery || !("data" in callbackQuery)) return;
+
       const match = callbackQuery.data?.match(/^resume_order_(.+)$/);
       if (!match) return;
 
       const orderId = match[1];
-      await this.volumeOrderService.updateOrderStatus(orderId, OrderStatus.RUNNING);
-      
-      await ctx.answerCbQuery('Order resumed');
+      await this.volumeOrderService.updateOrderStatus(
+        orderId,
+        OrderStatus.RUNNING,
+      );
+
+      await ctx.answerCbQuery("Order resumed");
       await this.handleViewOrder(ctx); // Refresh the view
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
@@ -1047,15 +1247,18 @@ ${order.completed_at ? `✨ **Completed:** ${new Date(order.completed_at).toLoca
   async handleStopOrder(ctx: Context): Promise<void> {
     try {
       const callbackQuery = ctx.callbackQuery;
-      if (!callbackQuery || !('data' in callbackQuery)) return;
-      
+      if (!callbackQuery || !("data" in callbackQuery)) return;
+
       const match = callbackQuery.data?.match(/^stop_order_(.+)$/);
       if (!match) return;
 
       const orderId = match[1];
-      await this.volumeOrderService.updateOrderStatus(orderId, OrderStatus.COMPLETED);
-      
-      await ctx.answerCbQuery('Order stopped');
+      await this.volumeOrderService.updateOrderStatus(
+        orderId,
+        OrderStatus.COMPLETED,
+      );
+
+      await ctx.answerCbQuery("Order stopped");
       await this.handleViewOrder(ctx); // Refresh the view
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
@@ -1068,22 +1271,22 @@ ${order.completed_at ? `✨ **Completed:** ${new Date(order.completed_at).toLoca
   async handleDeleteOrder(ctx: Context): Promise<void> {
     try {
       const callbackQuery = ctx.callbackQuery;
-      if (!callbackQuery || !('data' in callbackQuery)) return;
-      
+      if (!callbackQuery || !("data" in callbackQuery)) return;
+
       const match = callbackQuery.data?.match(/^delete_order_(.+)$/);
       if (!match) return;
 
       const orderId = match[1];
       const order = await this.volumeOrderService.getOrder(orderId);
-      
+
       if (!order) {
-        await ctx.answerCbQuery('Order not found');
+        await ctx.answerCbQuery("Order not found");
         return;
       }
 
       // Only allow deletion of pending payment orders
       if (order.status !== OrderStatus.PENDING_PAYMENT) {
-        await ctx.answerCbQuery('Can only delete unpaid orders');
+        await ctx.answerCbQuery("Can only delete unpaid orders");
         return;
       }
 
@@ -1097,13 +1300,16 @@ Are you sure you want to delete this order?
 This action cannot be undone.`;
 
       await ctx.editMessageText(message, {
-        parse_mode: 'Markdown',
+        parse_mode: "Markdown",
         reply_markup: Markup.inlineKeyboard([
           [
-            Markup.button.callback('✅ Yes, Delete', `confirm_delete_${orderId}`),
-            Markup.button.callback('❌ Cancel', `view_order_${orderId}`)
-          ]
-        ]).reply_markup
+            Markup.button.callback(
+              "✅ Yes, Delete",
+              `confirm_delete_${orderId}`,
+            ),
+            Markup.button.callback("❌ Cancel", `view_order_${orderId}`),
+          ],
+        ]).reply_markup,
       });
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
@@ -1116,19 +1322,19 @@ This action cannot be undone.`;
   async handleConfirmDelete(ctx: Context): Promise<void> {
     try {
       const callbackQuery = ctx.callbackQuery;
-      if (!callbackQuery || !('data' in callbackQuery)) return;
-      
+      if (!callbackQuery || !("data" in callbackQuery)) return;
+
       const match = callbackQuery.data?.match(/^confirm_delete_(.+)$/);
       if (!match) return;
 
       const orderId = match[1];
       await this.volumeOrderService.deleteOrder(orderId);
-      
-      await ctx.answerCbQuery('Order deleted');
-      await ctx.editMessageText('🗑️ Order deleted successfully.', {
+
+      await ctx.answerCbQuery("Order deleted");
+      await ctx.editMessageText("🗑️ Order deleted successfully.", {
         reply_markup: Markup.inlineKeyboard([
-          [Markup.button.callback('🔙 Back to Orders', 'orders_page_0')]
-        ]).reply_markup
+          [Markup.button.callback("🔙 Back to Orders", "orders_page_0")],
+        ]).reply_markup,
       });
     } catch (error) {
       await this.errorService.handleError(ctx, error, ErrorType.GENERAL, {});
